implement pr feedback
