+++ nodeEditor/src/diagram/graphCanvas.tsx
@@ -842,30 +842,30 @@ export class GraphCanvasComponent extends React.Component<IGraphCanvasComponentP
     }
 
     processEditorData(editorData: IEditorData, isImportingAFrame = false) {
-        if (!isImportingAFrame) {
-            const frames = this._frames.splice(0);
-            for (var frame of frames) {
-                frame.dispose();
-            }
-
-            this._frames = [];
-            this.x = editorData.x || 0;
-            this.y = editorData.y || 0;
-            this.zoom = editorData.zoom || 1;
+        const frames = this._frames.splice(0);
+        for (var frame of frames) {
+            frame.dispose();
         }
+
+        this._frames = [];
+        this.x = editorData.x || 0;
+        this.y = editorData.y || 0;
+        this.zoom = editorData.zoom || 1;
 
         // Frames
         if (editorData.frames) {
-            if (isImportingAFrame) {
-                const frame = GraphFrame.Parse(editorData.frames[editorData.frames.length - 1], this, editorData.map);
+            for (var frameData of editorData.frames) {
+                var frame = GraphFrame.Parse(frameData, this, editorData.map);
                 this._frames.push(frame);
-                this.globalState.onSelectionChangedObservable.notifyObservers(frame);
-            } else {
-                for (var frameData of editorData.frames) {
-                    var frame = GraphFrame.Parse(frameData, this, editorData.map);
-                    this._frames.push(frame);
-                }
-            }
+            }  
+        }
+    }
+
+    addFrame(editorData: IEditorData) {
+        if (editorData.frames) {
+            const frame = GraphFrame.Parse(editorData.frames[editorData.frames.length - 1], this, editorData.map);
+            this._frames.push(frame);
+            this.globalState.onSelectionChangedObservable.notifyObservers(frame);
         }
     }
