+++ nodeEditor/src/serializationTools.ts
@@ -52,7 +52,6 @@ export class SerializationTools {
     public static AddFrameToMaterial(serializationObject: any, globalState: GlobalState, currentMaterial: NodeMaterial) {
         globalState.onIsLoadingChanged.notifyObservers(true);
         this.UpdateLocations(currentMaterial, globalState);
-        globalState.nodeMaterial!.addFrameNodesFromeSerialization(serializationObject, "");
-        globalState.onImportFrameObservable.notifyObservers();
+        globalState.onImportFrameObservable.notifyObservers(serializationObject);
     }
 }
